* Fixed bug in HAT * little improvement of pagecached memory object * improve sys_memobj_create syscall

+++ include/mlibc/types.h
@@ -15,7 +15,7 @@ typedef uint_t bool;
 typedef ulong_t size_t;
 typedef uint16_t uid_t;
 typedef uint16_t gid_t;
-typedef uint64_t mode_t;
+typedef uint32_t mode_t;
 
 #ifdef CONFIG_ALWAYS_INLINE
 #define always_inline inline __attribute__((always_inline))
